update test function

+++ include/dft.h
@@ -3,6 +3,6 @@
 // discrete fourier transform
 void DFT(Mat& I);
 // standard deviations of frequencies
-void sdof(Mat& I, vector<float>& da, size_t& rings);
+void sdof(Mat& I, vector<float>& da, int& rings);
 // process dft and output features
-void process_dft(Mat& I, vector<float>& da, size_t& rings);
+void process_dft(Mat& I, vector<float>& da, int& rings);
